backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-pymongo==2.3.0
python-dotenv==1.0.0
pymongo==4.6.0
bcrypt==4.1.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
prophet==1.1.5
openai==1.12.0
httpx==0.26.0